Moved and renamed existing alerts.
