+++ src/components/editor/UnusedRDFDisplay.jsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect } from "react"
 import { useSelector } from "react-redux"
-import Alert from "../Alert"
+import Alert from "components/alerts/OldAlert"
 import RDFDisplay from "components/editor/preview/RDFDisplay"
 import { selectCurrentResourceKey } from "selectors/resources"
 import { selectUnusedRDF } from "selectors/modals"
