+++ src/components/editor/property/PropertyComponent.jsx
@@ -6,7 +6,7 @@ import PropTypes from "prop-types"
 import InputLiteralOrURI from "../inputs/InputLiteralOrURI"
 import NestedResource from "./NestedResource"
 import ReadOnlyInputLiteralOrURI from "../inputs/ReadOnlyInputLiteralOrURI"
-import Alert from "../../Alert"
+import Alert from "components/alerts/OldAlert"
 import { displayResourceValidations } from "selectors/errors"
 import { selectUri } from "selectors/resources"
