+++ src/components/search/QASearchResults.jsx
@@ -15,7 +15,7 @@ import { getTerm, getContextValues } from "utilities/QuestioningAuthority"
 import useRdfResource from "hooks/useRdfResource"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faCopy } from "@fortawesome/free-solid-svg-icons"
-import Alerts from "../Alerts"
+import Alerts from "components/alerts/OldAlerts"
 import { selectErrors } from "selectors/errors"
 import _ from "lodash"
 import { datasetFromN3 } from "utilities/Utilities"
